test: extract swebench tooling suite
